Update per PR comments
