+++ sdks/go/pkg/beam/runners/prism/internal/jobservices/job.go
@@ -38,6 +38,7 @@ import (
 	jobpb "github.com/apache/beam/sdks/v2/go/pkg/beam/model/jobmanagement_v1"
 	pipepb "github.com/apache/beam/sdks/v2/go/pkg/beam/model/pipeline_v1"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/urns"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/worker"
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
@@ -92,8 +93,8 @@ type Job struct {
 	// Logger for this job.
 	Logger *slog.Logger
 
-	metrics            metricsStore
-	WorkerPoolEndpoint string
+	metrics metricsStore
+	mw      *worker.MultiplexW
 }
 
 func (j *Job) ArtifactEndpoint() string {
@@ -198,4 +199,14 @@ func (j *Job) Failed(err error) {
 	j.failureErr = err
 	j.sendState(jobpb.JobState_FAILED)
 	j.CancelFn(fmt.Errorf("jobFailed %v: %w", j, err))
+}
+
+func (j *Job) MakeWorker(env string) *worker.W {
+	wk := j.mw.MakeWorker(j.String()+"_"+env, env)
+	wk.EnvPb = j.Pipeline.GetComponents().GetEnvironments()[env]
+	wk.PipelineOptions = j.PipelineOptions()
+	wk.JobKey = j.JobKey()
+	wk.ArtifactEndpoint = j.ArtifactEndpoint()
+
+	return wk
 }
